+++ README.md
@@ -1,85 +1,36 @@
 # Курс по символьному исполнению для анализа Go программ
 
-Репозиторий с домашними заданиями для магистер1. **Изучите общую структуру:**
-```bash
-ls internal/    # Посмотрите на модули интерпретатора
-```
+Репозиторий с домашними заданиями для магистерского курса "Символьное исполнение".
 
-2. **Выполняйте ДЗ1 (CFG анализ):**
-```bash
-cd homework1
-# Изучите README.md
-# Реализуйте TODO методы в ../internal/cfg/
-go run main.go   # Тестирование
-```
+## Цели курса
 
-3. **Затем ДЗ2 (Символьные выражения):**
-```bash
-cd ../homework2  
-# Изучите README.md  
-# Реализуйте TODO методы в ../internal/symbolic/, ../internal/translator/
-go run main.go   # Тестирование
-```
+По итогам прохождения курса вы получите навыки:
+- Построения SSA для Go программ
+- Создания системы символьных выражений
+- Трансляции символьных выражений в SMT-формулы
+- Использования Z3 solver для поиска ошибок в коде
+- Разработки статических анализаторов символьного исполнения для Go
+- Генерации тестовых случаев автоматически
 
-4. **Итоговый результат:**
-Полнофункциональная система символьного анализа Go кода в `internal/`символьному исполнению. 
-Целью курса является создание инструментов статического анализа Go кода с использованием методов символьного исполнения и SMT-решателей.
-
-## 🎯 Цели курса
-
-К концу курса студенты смогут:
-- Строить Control Flow Graph (CFG) для Go программ
-- Создавать системы символьных выражений
-- Транслировать символьные выражения в SMT-формулы
-- Использовать Z3 solver для поиска ошибок в коде
-- Разрабатывать статические анализаторы для Go
-- **Научиться генерировать тестовые случаи** автоматически
-
-## 📚 Структура курса
-
-## 🏗 Архитектура проекта
+## Архитектура проекта
 
 Проект разделён на **общую кодовую базу** и **домашние задания**:
 
 ### Общая кодовая база: `internal/`
-Все компоненты интерпретатора находятся в единой базе кода:
+Все компоненты интерпретатора находятся в единой кодовой базе в директории `internal/`
 
-```
-internal/
-├── cfg/                    # Анализ Control Flow Graph (ДЗ1)
-│   ├── types.go           # Определения CFG типов (с TODO)
-│   ├── builder.go         # Построение CFG из SSA (с TODO)  
-│   └── visualizer.go      # Визуализация CFG (с TODO)
-├── symbolic/               # Символьные выражения (ДЗ2)
-│   ├── types.go           # Базовые типы и интерфейсы
-│   └── expressions.go     # Реализации выражений (с TODO)
-├── translator/             # Трансляция в SMT (ДЗ2)
-│   ├── interface.go       # Интерфейс транслятора
-│   └── z3_translator.go   # Z3 транслятор (с TODO)
-└── ssa_converter/          # SSA -> Symbolic (ДЗ2)
-    └── converter.go       # Конвертация (с TODO)
-```
+### Домашние задания
 
-### Домашние задания: развитие общей базы
+**ДЗ_1 - построение SSA (`homework1/`):**
 
-**ДЗ1 - CFG анализ (`homework1/`):**
-- Развивает модуль `internal/cfg/`
-- Содержит примеры и демонстрацию
-
-**ДЗ2 - Символьные выражения (`homework2/`):**  
-- Развивает модули `internal/symbolic/`, `internal/translator/`, `internal/ssa_converter/`
-- Использует CFG из ДЗ1
-- Содержит примеры и демонстрацию
-   - Архитектура системы
-   - Обработка AST Go
-   - Генерация тестовых случаев
+**ДЗ_2 - Символьные выражения и построение SMT (`homework2/`):**
 
 ### Примеры и утилиты
 
 - **[examples/](examples/)** - Демонстрационные примеры работы с Z3
 - **[pkg/z3wrapper/](pkg/z3wrapper/)** - Обёртка для удобной работы с Z3 solver
 
-## 🚀 Быстрый старт
+## Быстрый старт
 
 ### Требования
 
@@ -128,7 +79,7 @@ make examples
 ```bash
 cd homework1
 # Изучите README.md
-# Реализуйте TODO методы в cfg/
+# Реализуйте TODO методы в ssa/
 go run main.go
 ```
 
@@ -140,28 +91,7 @@ cd ../homework2
 go run main.go
 ```
 
-## 📖 Теоретические материалы
-
-### Основные концепции
-
-- **Символьное исполнение** - метод анализа программ, использующий символьные значения вместо конкретных
-- **Path constraints** - ограничения, накапливаемые при прохождении по путям выполнения
-- **SMT solver** - инструмент для решения формул в теориях первого порядка
-- **Test case generation** - автоматическое создание тестовых данных
-
-### Рекомендуемая литература
-
-1. **"Introduction to Software Testing"** - Paul Ammann, Jeff Offutt
-2. **"The Art, Science, and Engineering of Fuzzing: A Survey"** - Valentin J.M. Manès et al.
-3. **"Symbolic Execution for Software Testing: Three Decades Later"** - Cristian Cadar, Koushik Sen
-
-### Полезные ссылки
-
-- [Z3 Theorem Prover](https://github.com/Z3Prover/z3)
-- [Go AST documentation](https://pkg.go.dev/go/ast)
-- [SAGE: Symbolic Execution Tool](https://www.microsoft.com/en-us/research/project/sage/)
-
-## 🛠️ Инструменты разработки
+## Инструменты разработки
 
 ### Makefile команды
 
@@ -170,45 +100,9 @@ make build          # Сборка всех примеров
 make test          # Запуск тестов
 make clean         # Очистка временных файлов
 make examples      # Запуск всех примеров
-make homework1     # Проверка домашнего задания 1
-make homework2     # Проверка домашнего задания 2
+make hw1     # Проверка домашнего задания 1
+make hw2     # Проверка домашнего задания 2
 ```
-
-### Структура проекта
-
-```
-symbolic-execution-course/
-├── internal/                   # 🏗 Общая кодовая база интерпретатора
-│   ├── cfg/                   # CFG анализ (ДЗ1)
-│   ├── symbolic/              # Символьные выражения (ДЗ2)
-│   └── translator/            # SMT трансляция (ДЗ2)
-├── homework1/                 # 📚 ДЗ1: CFG задания и примеры
-├── homework2/                 # 📚 ДЗ2: Символьные выражения, задания и примеры
-├── examples/                  # Демонстрационные примеры Z3
-│   └── basic_z3_example.go   # Основы работы с Z3
-├── pkg/
-│   └── z3wrapper/            # Обёртка для Z3
-├── go.mod                    # Go модуль с общими зависимостями
-├── Makefile                  # Автоматизация
-└── README.md                 # Этот файл
-```
-
-## 🎓 Система оценивания
-
-- **Домашнее задание 1**: 40 баллов
-- **Домашнее задание 2**: 50 баллов  
-- **Участие в семинарах**: 10 баллов
-
-**Итого**: 100 баллов
-
-### Критерии оценки
-
-- **Корректность реализации** (50%)
-- **Полнота тестирования** (20%)
-- **Качество кода и документации** (20%)
-- **Креативность и дополнительные возможности** (10%)
-
-## 🤝 Сотрудничество
 
 ### Правила выполнения заданий
 
@@ -216,39 +110,5 @@ symbolic-execution-course/
 - Обсуждение подходов разрешено, копирование кода - нет
 - Использование внешних библиотек должно быть согласовано
 - Все источники должны быть указаны в комментариях
-
-### Получение помощи
-
-- **Issues** в репозитории для вопросов по заданиям
-- **Discussions** для общих вопросов по курсу
-- **Office hours**: пт 14:00-16:00 (аудитория 123)
-
-## 📅 Расписание
-
-| Неделя | Тема | Задание |
-|--------|------|---------|
-| 1-2 | Введение в символьное исполнение | HW1 выдано |
-| 3 | Работа с Z3 solver | |  
-| 4 | Анализ путей выполнения | HW1 сдача, HW2 выдано |
-| 5-6 | Циклы и рекурсия | |
-| 7-8 | Структуры данных | HW2 сдача |
-| 9-10 | Архитектура символьных машин | |
-| 11 | Оптимизации и продвинутые техники | |
-| 12 | Защита проектов | Итоговая презентация |
-
-## 📊 Результаты обучения
-
-По завершении курса студенты смогут:
-
-✅ Объяснить принципы символьного исполнения  
-✅ Использовать Z3 для решения символьных ограничений  
-✅ Анализировать Go программы на предмет ошибок  
-✅ Проектировать архитектуру символьных анализаторов  
-✅ Генерировать тестовые случаи автоматически  
-✅ Оптимизировать процесс символьного анализа  
-
----
-
-**Удачи в изучении символьного исполнения! 🚀**
 
 *При возникновении вопросов обращайтесь к преподавателю или создавайте issue в репозитории.*
