+++ go.mod
@@ -3,6 +3,6 @@ module symbolic-execution-course
 go 1.23.2
 
 require (
-	github.com/aclements/go-z3 v0.0.0-20220809013456-4675d5f90ca5
+	github.com/ebukreev/go-z3 v0.0.0-20250821144348-dfd1fde1462b
 	golang.org/x/tools v0.13.0
 )
