+++ homework1/examples/test_functions.go
@@ -1,4 +1,4 @@
-// Простые тестовые функции для анализа CFG
+// Простые тестовые функции для построения SSA
 package main
 
 // Функция с простым условным оператором
