+++ homework2/examples/test_functions.go
@@ -1,4 +1,4 @@
-// Примеры функций для тестирования конвертации SSA в символьные выражения
+// Примеры функций для тестирования конвертации символьных выражений в SMT запрос
 package main
 
 // Простая арифметическая функция
