+++ internal/symbolic/types.go
@@ -24,101 +24,3 @@ func (et ExpressionType) String() string {
 		return "unknown"
 	}
 }
-
-// SymbolicExpression - базовый интерфейс для всех символьных выражений
-type SymbolicExpression interface {
-	// Type возвращает тип выражения
-	Type() ExpressionType
-
-	// String возвращает строковое представление выражения
-	String() string
-
-	// Accept принимает visitor для обхода дерева выражений
-	Accept(visitor Visitor) interface{}
-}
-
-// Visitor интерфейс для обхода символьных выражений (Visitor Pattern)
-type Visitor interface {
-	VisitVariable(expr *SymbolicVariable) interface{}
-	VisitIntConstant(expr *IntConstant) interface{}
-	VisitBoolConstant(expr *BoolConstant) interface{}
-	VisitBinaryOperation(expr *BinaryOperation) interface{}
-	VisitLogicalOperation(expr *LogicalOperation) interface{}
-	// TODO: Добавьте методы для других типов выражений по мере необходимости
-}
-
-// Операторы для бинарных выражений
-type BinaryOperator int
-
-const (
-	// Арифметические операторы
-	ADD BinaryOperator = iota
-	SUB
-	MUL
-	DIV
-	MOD
-
-	// Операторы сравнения
-	EQ // равно
-	NE // не равно
-	LT // меньше
-	LE // меньше или равно
-	GT // больше
-	GE // больше или равно
-)
-
-// String возвращает строковое представление оператора
-func (op BinaryOperator) String() string {
-	switch op {
-	case ADD:
-		return "+"
-	case SUB:
-		return "-"
-	case MUL:
-		return "*"
-	case DIV:
-		return "/"
-	case MOD:
-		return "%"
-	case EQ:
-		return "=="
-	case NE:
-		return "!="
-	case LT:
-		return "<"
-	case LE:
-		return "<="
-	case GT:
-		return ">"
-	case GE:
-		return ">="
-	default:
-		return "unknown"
-	}
-}
-
-// Логические операторы
-type LogicalOperator int
-
-const (
-	AND LogicalOperator = iota
-	OR
-	NOT
-	IMPLIES
-)
-
-// String возвращает строковое представление логического оператора
-func (op LogicalOperator) String() string {
-	switch op {
-	case AND:
-		return "&&"
-	case OR:
-		return "||"
-	case NOT:
-		return "!"
-	case IMPLIES:
-		return "=>"
-	default:
-		return "unknown"
-	}
-}
